+++ pest/src/lib.rs
@@ -4,13 +4,10 @@ mod inputs;
 mod parser;
 mod parser_state;
 mod error;
-mod token;
-mod token_stream;
+pub mod tokens;
 
 pub use inputs::Input;
 pub use inputs::StringInput;
 pub use parser::Parser;
 pub use parser_state::ParserState;
 pub use error::Error;
-pub use token::Token;
-pub use token_stream::TokenStream;
