--- pest/src/token.rs
@@ -1,8 +1,0 @@
-/// An `enum` representing tokens generated by a parser.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum Token<Rule> {
-    /// The starting bit of the `Token`.
-    Start(Rule, usize),
-    /// The ending bit of the `Token`.
-    End(Rule, usize)
-}
