+++ pest/src/tokens/token.rs
@@ -1,0 +1,14 @@
+/// An `enum` representing tokens generated by a parser.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub enum Token<Rule> {
+    /// The starting bit of the `Token`
+    Start {
+        rule: Rule,
+        pos:  usize
+    },
+    /// The ending bit of the `Token`
+    End {
+        rule: Rule,
+        pos:  usize
+    }
+}
