+++ pest/src/tokens/token_stream.rs
@@ -2,7 +2,7 @@ use futures::{Async, Poll};
 use futures::stream::Stream;
 use futures::sync::mpsc::UnboundedReceiver;
 
-use super::error::Error;
+use super::super::error::Error;
 use super::token::Token;
 
 /// A `Stream` of `Token` objects generated by the `Parser`.
@@ -19,7 +19,8 @@ impl<Rule> TokenStream<Rule> {
     /// # extern crate futures;
     /// # extern crate pest;
     /// # use futures::sync::mpsc::unbounded;
-    /// # use pest::{Error, Token, TokenStream};
+    /// # use pest::Error;
+    /// # use pest::tokens::{Token, TokenStream};
     /// # fn main() {
     /// let (_, r) = unbounded::<Result<Token<()>, Error<()>>>();
     ///
